Soporte para ventas por peso: selector de medida y ajuste en ticket
Se agregó un selector de tipo de medida (Unidad o Peso) en la gestión de productos, ajustando etiquetas de precio de compra y venta.\n\nEn el punto de venta se adaptó la estructura para soportar ventas por gramaje.\n\nAdemás, se añadió el tipo de medida al ticket para que se muestre correctamente en el detalle impreso.

+++ UsaloYa.Services/interfaces/IProductService.cs
@@ -25,7 +25,7 @@ namespace UsaloYa.Services.interfaces
         Task<Product4InventariotDto> SetInVentario(string barcode, int quantity, int companyId);
         Task<bool> SetAllUnitsStock(int companyId);
         Task<bool> ResetAllInVentario(int companyId);
-        Task<int> SetUnitsInStockByProductId(int productId, int companyId);
-        Task<int> SetUnitsInStock(int productId, int unitsInStock, bool isHardReset, int companyId);
+        Task<decimal> SetUnitsInStockByProductId(int productId, int companyId);
+        Task<decimal> SetUnitsInStock(int productId, decimal unitsInStock, bool isHardReset, int companyId);
     }
 }
